J_Money Scan/requirements.txt
gspread
oauth2client
PyYAML
python-telegram-bot
pandas
requests
dotenv
gspread
pytz